requirements.txt
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
streamlit==1.13.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
qrcode==7.3.1